perf: push account secret
